[SYCL] Added support for aspects and the has() function.
The SYCL 2020 Provisional Spec adds a new concept called device "aspects",
which provide a way for an application to query whether a device supports
certain features which are not available on all devices.

The device::has() function returns true if a device supports the aspect,
and false otherwise.  The platform::has() function returns true if all
devices on the platform support the given aspect, and false otherwise.

Signed-off-by: Gail Lyons <gail.lyons@intel.com>