[DAE][SYCL] Emit MD instead of updating integration header (#2258)
Removed DAE from standard optimization pipeline

We cannot run DAE in ESIMD cntext since the pointers to SPIR kernel
functions are saved in !genx.kernels metadata. Because of that I
removed DAE pass from the standard pipeline and put it in the place
that is specific for SYCL passes and guarded it under a new
temporary option: '-fenable-sycl-dae'. This option is only used until
other parts of the solution are not in place.