[SYCL] Wrap all pointers inside a generated struct
Commit 0b2de9e implemented decomposition of kernel parameters
of struct type i.e. fields of a struct are now individually
passed as separate kernel arguments. This results in an error
for pointer fields, violating SYCL spec. Copy of structs with
pointers is allowed according to the spec (value of pointer is
undefined). This patch fixes this error by wrapping all pointer
fields in a 'wrapper_struct'.

Signed-off-by: Elizabeth Andrews <elizabeth.andrews@intel.com>