[SYCL][WIP] Move kernel diagnostics to the call
This patch is a WIP that intends to move the diagnostics for a kernel to
the point in which it is called for the first time. This should give us
access to the actual call location (AND provide for diagnostics up an
instantiation tree) to provide bettter feedback to the user.

Currently, this doesn't actually add the call location anywhere (except
it moves the 'lambda' capture check to the call location, perhaps
incorrectly).

Additionally, this patch doesn't pass accessor-type-diagnostics for some
reason, so I still need to look into that.

Finally, this starts diagnosing that the kernel isn't a cxx-record-decl,
since that is in violation of the spec at the moment.

Also note there are a couple of TODOs I have to look into.