add option to set (i.e. reduce) build parallelism
the default option of "all the cores" or "4" is imprudent on some
systems, or when the user wants the system to be more responsive to
other applications.

this change was inspired by the thermal death (thankfully temporary)
of my Raspberry Pi, which has 4 cores but overheats when building LLVM
on all of them (a heatsink or fan would presumably resolve this).

Signed-off-by: Hammond, Jeff R <jeff.r.hammond@intel.com>