[SYCL] Move kernel diagnostics to the cal and, enable template trail on errors (#2289)
This patch moves the diagnostics for the kernel to the point in which the
kernel is called, rather than when generating the kernel. This means that
the call location is available (so that the checkers can be taught how to print
better diagnostics in the future). The biggest impact of this, is that it
causes us to print the template instantiation 'tree' since these are diagnosing
right away.

Currently, this uses the call location to better diagnose the lambda check (see
the additional note there!), and prints a kernel-parameter-must-be-lambda-or-functor
error.