[SYCL] Wrap pointer fields inside a generated struct (#2288)
Commit 0b2de9e implemented decomposition of kernel parameters
of struct type i.e. fields of a struct are now individually
passed as separate kernel arguments. This results in an error
for pointer fields of the struct, violating SYCL spec. Copy of structs
with pointers is allowed according to the spec (value of pointer is
undefined). This patch fixes this error by wrapping the pointer
fields in a '__wrapper_class' after the decomposition of struct type.

Signed-off-by: Elizabeth Andrews elizabeth.andrews@intel.com