[Buildbot] Add ARM target option to configure.py (#2322)
This has been tested on Raspberry Pi 4 and Cavium ThunderX2 running Ubuntu AArch64.  The build also depends on my fixes for #2304, but those are not ready.  Nonetheless, this fix is independent of that one and can be merged without side effects.

There is a clearly a more general solution here, but it seems imprudent to add architectures to the buildbot testing until they have actually been tested somewhere.  Once I have a chance to test on PowerPC, I'll see about adding that as an explicit target, or provide a more general solution (e.g. `--arch=..`).

Signed-off-by: Jeff R. Hammond <jeff.r.hammond@intel.com>