[NFC][SYCL] Simplify the VisitUnion functionality to be more reusable.
The VisitUnion implmentation that I gave Soumi to do this ended up
becoming way too complicated. This patch takes a slightly different
implementation tack in order to be significantly less code.

This patch uses a 'no-op' handler, one that just does nothing in order
to just replace the handlers that no longer need to visit, rather than
removing them from the list. This is slightly less efficient than the
old way, since there are now more function calls, but I don't think this
is a problem for 2 reasons:

1- The noop handler is marked 'final', AND is trivially inlinable, AND
does nothing, a moderately smart optimizer can just remove the calls.

2- Since we've broken up the diagnostic visitors into a separate step
(during call-expr creation, instead of all at once), my suspicion is
that it will be an 'all or nothing' sort of thing.