[SYCL] Immutable kernel functions redo. (#2259)
This is the same immutable kernel functions PR that was merged (and reverted) last week. The SYCL conformance tests have been updated, so hopefully this should be mergeable.   I've added the `__SYCL_NONCONST_FUNCTOR__` as requested.

I have also hooked up the `-sycl-std=` compiler driver flag so that it will pass `DCL_SYCL_LANGUAGE_VERSION=121` if SYCL 1.2.1 is chosen or `DCL_SYCL_LANGUAGE_VERSION=2020` if SYCL 2020 is chosen.

In SYCL 2020 kernel functions are now const and passed by reference.  In this PR the existing API constant CL_SYCL_LANGUAGE_VERSION now determine another constant ( __SYCL_NONCONST_FUNCTOR__ ) that makes the kernel functions be constant references (or not). 

In the Clang Front end, both by-reference and by-value passing of the kernel functions is supported, and a diagnostic emitted if that is mismatched.  The `sycl-std` flag now supports `2020` as a valid option.  And the `sycl-std` flag will pass `-DCL_SYCL_LANGUAGE_VERSION=` to match its election.    Tests have been added to confirm that both kernel passings are supported and that the sycl-std flag is working correctly.   Additionally, there are a lot of FE tests that employ mock kernels, these have been updated to the SYCL 2020 convention avoiding the deferred maintenance.  

Signed-off-by: Chris Perkins <chris.perkins@intel.com>