[NFC][SYCL] Remove enterField/leaveField from int-header
In an effort to clean up the visitor/redo some array handling, I
noticed that the enter-field/leave-field mechanism is inconsistent with
the design of the visitor.  This patch starts the effort to remove it
entirely, and move the field-based dependence BACK to the individual
fields as required.

This DID require adding a field to the special visitor types for bases,
which generally seems to make these functions completely unnecessary.