[NFCI][SYCL] WIP On Rework of how arrays function.
This is a WIP patch, just for information of Mariya/Elizabeth to see how
things are going.

The intent of this patch is simply to clean up how arrays work in the
SYCL visitors.  This patch already changes the visitor in a way that I
believe is optimal and has what is necessary to finish this.

Integration header generation also seems to be correct.

Work on this DID require quite a bit of a re-imagining of how the
DeclBodyCreator works. There are a number of small refactors to fix
things I found while working that I believe make it easier to maintain
this in the future.

However, the biggest change here is with the init-list-expr creation.
This patch creates them in advance, rather than waiting until we are
done with a struct/array/etc.  This allows us to generate them as we go.

The side effect of this is that the InitExprs array now ONLY contains
collection InitListExprs, since it no longer has to assemble them on the
other side.

TODO List:
There are a number of TODOs in the patch itself that should be worked
out at one point or another.  There are also currently 9 failing lit
tests.  However, I know that the following doesn't work:
- Base Class initialization lists
- Base Class MemberExprBases
- Array MemberExprBases