Do not use std::set in getRequiredExtensions
In all cases what we have encountered so far each instruction has either
zero or one extension that it requires and therefore there is no need to
use `std::set` to store that information.

Replaced `std::set` with `llvm::Optional` and renamed function to
indicate that it can only return a single extension.