[SYCL] Make host accessor commands bypass leaf limit (#2321)
Extend functionality of CircularBuffer class to allow for a different way of
processing of host-accessors leaves. From now on host-accessors won't be
affected by leaf-limiting and, hence, they won't be removed from leaves list.

The new LeavesCollection class stores two distinct data structures:

    CircularBuffer<Command *> for generic commands; list of pointers to
    EmptyCommand (which is the blocked node of host-accessor) and map of xrefs
    to the list.

Map of xrefs is merely a mapping of EmptyCommand's pointer to iterator in the
list. This mapping allows for quick lookup if the command is there (in the
list).

The introduced class also provides forward-iterator fetched with
begin()/cbegin() and end()/cend() methods along with ability to fetch all the
leaves stored as a vector of pointers to Command.

Signed-off-by: Sergey Kanaev <sergey.kanaev@intel.com>