[SYCL] Enable L0 online link under envvar
Add temporary support to enable online linking with the Level Zero
driver if the SYCL_ENABLE_LEVEL_ZERO_LINK environment variable is
set.

At the time this code was written, the Level Zero driver online
linking APIs exist, but they don't work.  We think the DPC++ runtime
support is ready, though, and we can enable it via this environment
variable.  This will allow testing with a new driver if one is
available before the next DPC++ release.