[SYCL] Store stream buffers in the scheduler
Stream buffers need to be alive after submitting a kernel because it is
executed by the scheduler asynchronosly. For this reason currently
stream buffers are stored in an associated stream object. This stream
object is passed to the handler and then forwarded further to a commandi
group to keep streamm buffers alive for the scheduler.

But there is a problem with this approach. A command group cannot be
destroyed while stream buffers (which are accessed in this command
group) are alive. Stream buffers are destroyed only if the stream
is destroyed. Stream object is destrtoyed only if command group is
destroyed. So, there is a loop dependcy. Which results in memory leaks.

Solution is to store stream buffers in the scheduler for each stream.
With this approach resources are released properly.