[SYCL] Source line info for kernel initialization code (#2428)
Assigning the source location of the kernel caller function to the
artificial initialization code generated in the kernel body.

This change is to enable the profiling tool to meaningfully
attribute the initialization code. Before this change, the code
was attributed to 'line 0' of the file where the kernel object
is at, which the tools consider meaningless.