[SYCL] USM shared memory allocator for L0 plugin (#2366)
* [SYCL] USM shared memory allocator for L0 plugin

In L0 each allocation results in at least a memory page
no matter how small the size was requested. This adds
significant overhead when an app does many small allocations.

This patch adds a memory allocator on top of L0 USM API in
order to solve the problem.

High level description:
There is a predefine list of bucket sizes. When allocation function
is called, the best fitted bucket is found for the requested size.
If there is a free chunk in the bucket then it's returned, otherwise
a new slab of size 64k requested from the system via L0 API.
This slab is split into chunks of the size of corresponding bucket.
Now there is a free chunk in the slab which is returned.
When a chunk is returned it's marked as non-free until the user
call free with a corresponding pointer.

In addition to that add an environment variable
SYCL_PI_LEVEL0_DISABLE_USM_ALLOCATOR which disables the allocator
and returns back to the original behavior.

* [SYCL] Restrict USM allocators tests to Level zero

There are issues with running these tests on OpenCL
on windows.
The primary goal of these tests is to test USM allocator
which is currently emabled only for Level zero, so disable
them on OpenCL for now.

* [SYCL] Fix pointer alignment

Due to incorrect pointer alignment in case when
allcation alignment > size we returned to the user
and incorrect chunk(i.e. next to the one which was
marked as non-free) which is later detected as "double free"
error.

Also corrected the existing testcases to test this scenario