[SYCL] Diagnose attempt to pass pointer to VLA as kernel arg (#2441)
VLA as well as pointers to VLA require additional AST transformation to make
emission of corresponding type in LLVM IR possible. Without this transformation
Codegen just crashes.
Implementing this transformation is not reasonable whereas SYCL standard
doesn't allow VLA in device code, so emit an error in case if pointer to
VLA is passed as kernel argument.