Move the decomp checker to its own invocation of the visitor
This is necessary since the size-checker needs opt-in (so that it
properly reflects the opencl kernel arguments).

When they are in the same invocation, the size-checker is erronously
called thinking that the first time we see a struct that it doesn't need
to be decomposed.

I opted to use the same visitor, since it doesn't have state.