[SYCL] Fixed bugs regarding Platform Caching
Added new Platforms that have been created and initialized to the
PiPlatformCache.

Created a new function, getOrCreatePlatform, so that the cached Platforms could be
shared accross both piPlatformsGet and piextPlatformCreateWithNativeHandle.

Handle platform initialization in piCachePlatforms, so that platforms created
with a native handle would also be initialized (Before initialization was not
occuring for platforms created in piextPlatformCreateWithNativeHandle).