[SYCL] Remove kernel_signature_start from int header
Some time ago, the `kernel_signatures` array used `kind_none` as a delimiter between signatures of different kernels. The `kernel_signature_start` array's values were computed with a `+ 1` to account for this. Then one day the delimiter was removed but the `+ 1` remained. It wasn't noticed that the values were off by 1 because `kernel_signature_start` isn't being used anymore anyway. Therefore in this change, we simply remove it.

GH issue #2490