[SYCL][ESIMD] Fixed compiler crash in LowerESIMDVecArg pass
This fixes potential compiler crash in LowerESIMDVecArg pass, which I encountered while writing a small test. Just to be clear, this doesn't happen in a real test, but potentially could happen.

The problem arises when Global is used in simple instruction, not directly in ConstantExpr, e.g.:

```
@GlobalGRF_data = dso_local global %"class._ZTSN2cl4sycl5INTEL3gpu4simdIiLi2512EEE.cl::sycl::INTEL::gpu::simd" undef, align 16384

define void @no_crash(<2512 x i32> %simd_val) {
  %cast = addrspacecast %"class._ZTSN2cl4sycl5INTEL3gpu4simdIiLi2512EEE.cl::sycl::INTEL::gpu::simd"* @GlobalGRF_data to %"class._ZTSN2cl4sycl5INTEL3gpu4simdIiLi2512EEE.cl::sycl::INTEL::gpu::simd" addrspace(4)*
  ...
}
```

It crashed in `ESIMDLowerVecArgPass::createNewConstantExpr`.