[SYCL] Release commands with no dependencies after they're enqueued (#2492)
Presently, commands that do not have memory dependencies are only released if .wait() is called. If it is not, several resources (event, queue, command, kernel) are held onto.
This fix deletes stores that command event in the USMEvents ( so it is 'owned') and then deletes the command itself. Added a lit test to verify resource release.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>