[SYCL] Relax test to work in Win32 environment.
Our downstream runs this test on a 32 bit windwos machine, so relax the
test slightly to allow it to pass there.

The problem is that the type of 'size_t' (the array index type) is
different (it is unsigned int instead of unsigned long/unsigned long
long), so this changes the test to allow all 3.