Rename OCL20ToSPIRV pass
Renamed the pass to OCLToSPIRV: the problem with the old name is that it
doesn't represent actual pass content: it handles not only OpenCL C 2.0
built-in functions, but also OpenCL C 1.2 built-in functions, acting
like an adapter between OpenCL C representation and SPIR-V friendly IR
representation. Also updated some comments to reflect that.