[SYCL][L0] Fixed SYCL_PI_TRACE printing problem
SYCL_PI_TRACE is used for internal debugging. It displays
the PI function that is being called and its arguments.

When piGetDeviceInfo() was called, the printing of the
pointer used for a variable sized string was incorrect.
Added a print() specialization.

The nullptr print() specialization was also updated because
specializations should be special.

Signed-off-by: Gail Lyons <gail.lyons@intel.com>