[SYCL] Wrap complex global objects to control lifetime (#2516)
The order of global constructor/destructor invocation is undefined. At the same
time, user may desire to make objects of SYCL classes global. Since there're
lots of global objects inside SYCL runtime (scheduler and program manager, to
name a few), they need to be wrapped with some trivial class/structure. The
object of that structure is allocated on heap and deallocated in
__attribute__((destructor))/DllMain, which guarantees, that SYCL global objects
will be destroyed after user objects.