[SYCL] Fixed bug regarding device caching
Created a new function, populateDeviceCacheIfNeeded, so that cached devices could be
shared across both piDevicesGet and piextDeviceCreateWithNativeHandle. This
new function will check/fill and return cached devices.

Removed the function getOrCreatePlatforms.

Refactored PiPlatformsGet to limit the number Level Zero driver calls.

Called PiPlatformsGet from piextPlatformCreateWithNativeHandle so that the
latter has access to the platform cache maintained in the former.

Called getOrMakePlatformImpl from make_platform so that the PlatformImpl cache
is accessible to platforms that are created using a native handle.

Called getOrMakeDeviceImpl from make_device so that the DeviceImpl cache is
accessible to devices that are created using a native handle.

Also, added an E2E test for these changes.