[SYCL] Wait before deleting a stream buffer
Currently streamed data is printed in a host task. Host task has a
corresponding empty task which is a leaf command for a stream buffer.
Since stream buffer is not a leaf for a kernel it doesn't block the
kernel from being cleaned up after completion. During this cleanup
process stream buffer is deleted, as a result scheduler does a blocking
enqueue of the empty task. Problem is that this empty task has a
blocked status (blocked by host task). This results in exception saying
that we try to wait for a blocked command. It is not clear if this use
case is invalid or there is a bug in the scheduler.

Explicitly waiting for a host task before stream buffer deletion is
correct for sure and resolves this problem.