do not assume Linux always has aligned_alloc
Older glibc lacks aligned_alloc (C11 and C++17).
Dropping through to the POSIX option (posix_memalign) is necessary
on such systems.

The lazy way to detect if aligned_alloc is supported is:
defined(_GLIBCXX_HAVE_ALIGNED_ALLOC) || defined(_LIBCPP_HAS_C11_FEATURES)

This does not work for C++ libraries besides GCC and LLVM (e.g. MUSL)
but unless those are explicitly supported and tested, we should not
need to include them here anyways.

The inclusion of stdlib.h was changed to cstdlib, as appropriate for C++
code.  The conditional inclusion of cstdlib in the cpp file was removed
since it was included unconditionally in the header.

Signed-off-by: Jeff Hammond <jeff.r.hammond@intel.com>