[SYCL] Fix rectangle argument order when passing to PI routines (#2608)
For 2D and 3D buffers and images, some of the arguments passed to the plugin
interface are not always correct or in the right order. For example, with 2D buffer
we pass height-in-bytes, rather than width-in-bytes as part of a region or offset
argument. This fixes this, slightly expands the PI Trace, and and also adds testing
so we can more easily observe and catch this. Note that presently, images and
buffers use multidimensional range/ids in the reverse order of each other. This
fix does not attempt to unify those representations, but comments have been
added to help clarify this in the future.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>