[SYCL] Avoid including std C++ headers into integration header file (#2662)
This is done to avoid compilation errors caused by usage TBB AND libstdc++ 10.X

libstdc++10 conditionally defines (i.e. if is not defined already) the macro _GLIBCXX_USE_TBB_PAR_BACKEND.
If the integration header file has includes of C++ headers, then that macro gets defined always. This does not let to define it properly in user's app (even if define is on the 1st line), which also causes different behavior for some C++ std files and compile error later.

Signed-off-by: Vyacheslav N Klochkov <vyacheslav.n.klochkov@intel.com>