[SYCL][PI][L0] ZeSerialize false race fix (#2672)
Intel Inspector race condition detector doesn't think there is a race that
can occur when writing ZeSerialize on one thread while reading it on another.
There was never really a race before because all threads would write the same
value, and thus there was no issue.  But this way of writing the code removes
the false race, and makes it obvious this global variable can only be
initialized once.