[SYCL][LIT] Fix regex in select_device LIT test
This patch adds two new special characters - ```[``` and ```]``` to
select_device LIT test. An escape symbol will be added to these
characters. This is because new OpenCL GPU runtime 20.42.18209 has
such symbols in its device name, and regex generated by this test
was w/o escape symbols for ```[``` and ```]```, which led to
```CL_DEVICE_NOT_FOUND``` error in run time.