Revert "[SYCL][Driver] Propagate floating point optimizations flag (#2680)"
This reverts commit ce0888998126e62a97b21c576aaffd0c81db6cb5.

Reason: the issue with this implementation is that we can't preserve
correct correlation between the option and the source impacted by this
option. `-ffast-math` is applied to C++ translation unit, whereas single
SPIR-V module might include code sections from multiple translation
units and some of them might be compiled with `-fno-fast-math`.