[SYCL] Add support for new FPGA loop attribute nofusion
This patch adds support a new loop attribute for FPGA, intel::nofusion.
This attribute should be passed to the FPGA backend, and ignored by the emulator.
The attribute indicates that the annotated loop should not be fused with any adjacent loop.

Note: this does not include a corresponding [[intel::fusion]] attribute,
because a different mechanism (loop_fuse) will be built for FPGA.

Syntax:
[[intel::nofusion]]

The LLVM IR representation should be similar to the representation used for #pragma nofusion.
The llvm.loop metadata should specify llvm.loop.fusion.disable.

Signed-off-by: Soumi Manna <soumi.manna@intel.com>