[SYCL][PI][L0] Fix a problem with kernels and programs destruction while they can be used (#2710)
Requirements from summary of level zero spec:

1 – cannot do zeKernelDestroy until all commands using the kernel have completed
execution as measured by their event or fence being signaled.

2 – The application must destroy all kernel and build log handles created from
the module before destroying the module itself (zeModuleDestroy)
So, here is a design I think will work.

1 – We want to do a piKernelRetain sometime before kernel is appended to
command-list in piEnqueueKernelLaunch. The intent is that we will do a
piKernelRelease once the Event that we create to be associated with the Enqueued
Kernel is signaled/released. The handle of kernel to be released will be stored
in the event.

2 – Event currently contains fields for CommandType and CommandData, and for the
CommandType that piEnqueueKernelLaunch uses, CommandData is currently unused.
So, for this case, we will set CommandData to be the Kernel, so then when the
event is signaled as done, then we can use the kernel we stored in CommandData,
we can do piKernelRelease(CommandData).

3 – I think in piKernelRetain, we also need to call piProgramRetain on the
program that the kernel is part of, and we need piKernelRelease to call
piProgramRelease. This is necessary to make sure we do not destroy the program
before destroying all the kernels that may be in use from the program.

4 – in the destructor for pi_program, we need to destroy the build log before we
destroy the program itself.