[SYCL] Add support for new FPGA function attribute stall_enable
This patch adds support a new FPGA function attribute, stall_enable,
to be sent through to the backend (and ignored by the emulator).

Syntax:
[[intel::stall_enable]]

This function attribute applies to a lambda function, or function definition.
Requests, to the extent possible, that statically-scheduled clusters handle
stalls using a stall-enable signal to freeze computation within the cluster.

Note that this attribute can be applied to any lambda or function definition,
it does not have to be a device function/kernel.

Stall_enable attribute takes no arguments and LLVM IR function metadata
should simply be i32 1.

Signed-off-by: Soumi Manna <soumi.manna@intel.com>