[SYCL] Interoperability API to create a context from a Level Zero handler
Implement the missing API to create a SYCL context from a Level Zero
handler. Also add a test to check the interop API for the Level Zero.
This commit also includes changes to properly deploy the Level Zero
loader and the Level Zero headers during build/install in both
scenarios:
* The loader and headers are downloaded from github
* Paths to the loader and headers are provided to cmake

Signed-off-by: Artur Gainullin <artur.gainullin@intel.com>