[Driver][SYCL][FPGA] Allow for multiple AOCX device binaries for FPGA
Initial implementation only allowed for a single AOCR/AOCX binary to
be used in a single FPGA compilation.  This restriction is being lifted
and we are allowing multiple FPGA device images to be added to the
final executable.

Allow for list file creation for aocx/aocr files with update extensions
when unbundling the archives.  To handle this, we are also enabling the
rename functionality for file-table-tform