Added tests to verify that devicelib APIs have no 'noinline' attribute.
Device compilers may (and should) honor 'noinline' attribute
even when optimizations are turned on. At the same time, we do not want
to prevent inlining of devicelib wrappers and fallback implementations,
since this may result in degraded performance. These tests verify that
the devicelib APIs are not marked with 'noinline' attribute.

Signed-off-by: Vyacheslav Zakharin <vyacheslav.p.zakharin@intel.com>