[SYCL] Replace CodeGen changes with SPIR target info hooks
There are a lot of customizations in CodeGen library enabling generic
address space as default for SYCL mode. Based on feedback from the
community there is a better way to get similar results. CodeGen
TargetInfo object is be used to set default address space for global
variables and stack allocations.