[Driver][SYCL] Remove partial-link path when dealing with fat static archives
For Linux, we would use partial linking to generate the object that contained
the device binaries from a fat static archive.  We are moving away from
using partial linking to provide a better overall solution that is the
same in regards to link behaviors for both Linux and Windows.

The change here will allow for the extraction of device binaries from
archives occur in the same manner as it does for Windows.  We do full
extraction from the archives, and do a full llvm-link with those extracted
device binaries.