[SYCL] Optimize discard_write access mode (#2854)
When using discard_write with a host accessor, we do not need to keep the buffer updated with the latest changes on the device. Skipping that operation results in a significant speed-up.

When we have host unified memory, then Map/Unmap operations are enqueued. As these must always match, it is up to the backend (or its plugin) to take advantage of any optimization. In the case of Level Zero, this can be be done in the SYCL plugin interface code. This PR includes a fix for LevelZero.
CUDA requires no change. OpenCL will need to be changed in the library itself.

When we are not using host unified memory, instead of Map/Unmap operations, we enqueue basic Read/Write ops. When using discard_write on the host, the Read op is unnecessary. This patch schedules an empty operation instead.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>