[SYCL][L0] Fix memory leak for piCommandListFenceMap.
std::map::[]operator should not be used to lookup the existence of the map entry because it creates a place-holder entry if the entry does not exist.
This leads to a memory leak because the place-holder won't be explictly deallocated.
Instead, we should use std::map::find() to check if a map entry exists or not.

Signed-off-by: Byoungro So <byoungro.so@intel.com>