[SYCL] Adjust parallel-for range global size to improve group size selection (#2703)
This change rounds up a parallel-for range to be a multiple of 32. This value can be changed later when we have better strategies for selecting work-group sizes. It works well for now. The rounding-up improves performance by 8-10x for the odd cases when the original range is a prime number. It has negligible performance impact cases where the range is already a multiple of 32.

Signed-off-by: rdeodhar rajiv.deodhar@intel.com